/LinterAnalysis response body formatting. Codebase reorganization.
